requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
cloudinary>=1.40.0
python-telegram-bot>=21.0
requests>=2.31.0
